+++ fluree-db-policy/src/index.rs
@@ -215,7 +215,7 @@ mod tests {
     use crate::types::PolicyValue;
     use fluree_db_core::{ClassPropertyUsage, ClassStatEntry};
 
-    fn make_sid(ns: i32, name: &str) -> Sid {
+    fn make_sid(ns: u16, name: &str) -> Sid {
         Sid::new(ns, name)
     }
