+++ fluree-db-policy/src/schema.rs
@@ -83,7 +83,7 @@ pub fn is_schema_flake(predicate: &Sid, object: &FlakeValue) -> bool {
 mod tests {
     use super::*;
 
-    fn make_sid(ns: i32, name: &str) -> Sid {
+    fn make_sid(ns: u16, name: &str) -> Sid {
         Sid::new(ns, name)
     }
