+++ fluree-db-transact/src/generate/cancellation.rs
@@ -51,7 +51,7 @@ mod tests {
     use super::*;
     use fluree_db_core::{FlakeValue, Sid};
 
-    fn make_flake(s: i32, p: i32, o: i64, t: i64, op: bool) -> Flake {
+    fn make_flake(s: u16, p: u16, o: i64, t: i64, op: bool) -> Flake {
         Flake::new(
             Sid::new(s, format!("s{}", s)),
             Sid::new(p, format!("p{}", p)),
